+++ src/module/data/actor/base.mjs
@@ -109,7 +109,12 @@ export default class BaseActorModel extends DrawSteelSystemModel {
       dsid: new Set(),
     };
 
-    this.stamina.min = 0;
+    Object.assign(this.stamina, {
+      min: 0,
+      bonuses: {
+        echelon: 0,
+      },
+    });
 
     // Teleport speeds are unaffected by conditions and effects
     this.movement.teleport = this.movement.types.has("teleport") ? this.movement.value : null;
@@ -120,6 +125,9 @@ export default class BaseActorModel extends DrawSteelSystemModel {
   /** @inheritdoc */
   prepareDerivedData() {
     super.prepareDerivedData();
+
+    // Apply all stamina bonuses before calculating winded
+    this.stamina.max += this.echelon * this.stamina.bonuses.echelon;
 
     this.stamina.winded = Math.floor(this.stamina.max / 2);